Add : addOne function & Fix : update & delete

+++ controllers/userController.js
@@ -12,6 +12,75 @@ const getAll = async (res) => {
   }
 };
 
+const addOne = async (req, res) => {
+  if (!req.body.firstName) {
+    res.status(400).send({ error: "First name missing" });
+    return false;
+  }
+
+  if (!req.body.lastName) {
+    res.status(400).send({ error: "Last name missing" });
+    return false;
+  }
+
+  if (!req.body.userName) {
+    res.status(400).send({ error: "User name missing" });
+    return false;
+  }
+
+  if (!req.body.email) {
+    res.status(400).send({ error: "Email missing" });
+    return false;
+  }
+
+  if (!req.body.password) {
+    res.status(400).send({ error: "Password missing" });
+    return false;
+  }
+
+  if (!req.body.phone) {
+    res.status(400).send({ error: "Phone number missing" });
+    return false;
+  }
+
+  if (!req.body.gender) {
+    res.status(400).send({ error: "Gender missing" });
+    return false;
+  }
+
+  if (!req.body.birthday) {
+    res.status(400).send({ error: "Birthday missing" });
+    return false;
+  }
+
+  try {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: req.body.email,
+      },
+    });
+    if (user) {
+      res.status(400).send({ error: "Email already registered." });
+    } else {
+      const newUser = await prisma.user.create({
+        data: {
+          firstName: req.body.firstName,
+          lastName: req.body.lastName,
+          userName: req.body.userName,
+          email: req.body.email,
+          password: bcrypt.hashSync(req.body.password, 10),
+          gender: req.body.gender,
+          birthday: new Date(req.body.birthday),
+          phone: req.body.phone,
+        },
+      });
+      res.status(200).send({ success: "User Created!" });
+    }
+  } catch (err) {
+    res.status(400).send({ error: err });
+  }
+};
+
 const getOne = async (req, res) => {
   try {
     const user = await prisma.user.findUnique({
@@ -22,7 +91,7 @@ const getOne = async (req, res) => {
     if (user) {
       res.status(200).send({ user });
     } else {
-      res.status(400).send({ error: "Use not found!" });
+      res.status(400).send({ error: "User not found!" });
     }
   } catch (err) {
     console.log(err);
@@ -32,12 +101,22 @@ const getOne = async (req, res) => {
 
 const deleteOne = async (req, res) => {
   try {
-    const deleteUser = await prisma.user.delete({
+    const find = await prisma.user.findUnique({
       where: {
         id: parseInt(req.params.id),
       },
     });
-    res.status(200).send({ success: `User ${req.params.id} deleted!` });
+
+    if (find) {
+      const deleteUser = await prisma.user.delete({
+        where: {
+          id: parseInt(req.params.id),
+        },
+      });
+      res.status(200).send({ success: `User ${req.params.id} deleted!` });
+    } else {
+      res.status(400).send({ error: "User not found!" });
+    }
   } catch (err) {
     res.status(400).send({ error: err });
   }
@@ -45,19 +124,29 @@ const deleteOne = async (req, res) => {
 
 const updateOne = async (req, res) => {
   try {
-    const updateUsers = await prisma.user.updateMany({
+    const find = await prisma.user.findUnique({
       where: {
         id: parseInt(req.params.id),
       },
-      data: req.body,
     });
-    res
-      .status(200)
-      .send({ success: `User ${req.params.id} updated!`, user: updateUsers });
+
+    if (find) {
+      const updateUsers = await prisma.user.updateMany({
+        where: {
+          id: parseInt(req.params.id),
+        },
+        data: req.body,
+      });
+      res
+        .status(200)
+        .send({ success: `User ${req.params.id} updated!`, user: updateUsers });
+    } else {
+      res.status(400).send({ error: "User not found!" });
+    }
   } catch (err) {
     console.log(err);
     res.status(400).send({ error: err });
   }
 };
 
-module.exports = { getAll, getOne, deleteOne, updateOne };
+module.exports = { getAll, addOne, getOne, deleteOne, updateOne };
